requirements.txt
adjustText==0.7.3
amqp==2.5.2
appnope==0.1.0
asgiref==3.2.3
backcall==0.1.0
billiard==3.6.3.0
celery==4.4.1
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
deap==1.3.1
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
Django==3.0.7
django-allauth==0.41.0
django-celery-results==1.2.1
django-crispy-forms==1.9.0
django-custom-user==0.7
idna==2.9
importlib-metadata==1.5.0
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
joblib==0.14.1
kiwisolver==1.1.0
kombu==4.6.8
matplotlib==3.1.3
multiprocess==0.70.9
numpy==1.18.1
oauthlib==3.1.0
pandas==1.0.1
parso==0.6.2
pathos==0.2.5
pexpect==4.8.0
pickleshare==0.7.5
plotly==4.5.2
pox==0.2.7
ppft==1.6.6.1
prompt-toolkit==3.0.3
psycopg2==2.8.4
ptyprocess==0.6.0
pyDOE==0.3.8
pydream==1.1.0
Pygments==2.5.2
pyparsing==2.4.6
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
scikit-learn==0.22.2
scipy==1.4.1
sentry-sdk==0.14.2
six==1.14.0
sqlparse==0.3.0
swot==1.0.1
traitlets==4.3.3
uncertainties==3.1.2
urllib3==1.25.8
uWSGI==2.0.18
vine==1.3.0
wcwidth==0.1.8
zipp==3.1.0
